Performance increase

+++ src/engines/utils.rs
@@ -1,6 +1,6 @@
 /// Calculate the greatest common divisor of a list of numbers. Works by iterating through the list
 /// and testing the current gcd with the next number.
-pub fn gcd(values: Vec<u64>) -> u64 {
+pub fn gcd(values: Vec<usize>) -> usize {
     let mut gcd = values[0];
     for value in values {
         gcd = gcd_inner(gcd, value);
@@ -9,7 +9,7 @@ pub fn gcd(values: Vec<u64>) -> u64 {
 }
 
 /// Calculate the greatest common divisor of two numbers using Euclid's algorithm.
-fn gcd_inner(a: u64, b: u64) -> u64 {
+fn gcd_inner(a: usize, b: usize) -> usize {
     if b == 0 {
         a
     } else {
